==> recce1/6-simsmerge.r <==
### Analyze ironX and DISTANCE sampling
## rm(list = ls())
setwd("/Users/solson/Rprojects/recce1")
load("/Users/solson/Rprojects/recce1/simn5e5/k100.RData")
hk100<-k100
load("/Users/solson/Rprojects/recce1/simn5e5/k20.RData")
hk20<-k20
load("/Users/solson/Rprojects/recce1/simn5e5/k10.RData")
hk10<-k10
load("/Users/solson/Rprojects/recce1/simn5e5/k5.RData")
hk5<-k5
load("/Users/solson/Rprojects/recce1/simn5e5/k3IXs.RData")
load("/Users/solson/Rprojects/recce1/simn5e5/k3Dp1.RData")
load("/Users/solson/Rprojects/recce1/simn5e5/k3Dp2.RData")
load("/Users/solson/Rprojects/recce1/simn5e5/k3Dp3.RData")
hk3<-vector("list",10)
hk3[[8]]<-rbind(k3Dp1[[8]],k3Dp2[[8]],k3Dp3[[8]])
hk3[[9]][1:5]<-k3Dp1[[9]][1:5]
hk3[[9]][6:10]<-k3Dp2[[9]][1:5]
hk3[[9]][11:15]<-k3Dp3[[9]][1:5]
hk3[[10]][1:5]<-k3Dp1[[10]][1:5]
hk3[[10]][6:10]<-k3Dp2[[10]][1:5]
hk3[[10]][11:15]<-k3Dp3[[10]][1:5]
hk3[1:7]<-k3IXs[1:7]
rm(k100,k20,k10,k5,k3Dp1,k3Dp2,k3Dp3,k3IXs)




load("/Users/solson/Rprojects/recce1/simn3e5/k100IXs.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k100Dp1.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k100Dp2.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k100Dp3.RData")
mk100<-vector("list",10)
mk100[[8]]<-rbind(k100Dp1[[8]],k100Dp2[[8]],k100Dp3[[8]])
mk100[[9]][1:5]<-k100Dp1[[9]][1:5]
mk100[[9]][6:10]<-k100Dp2[[9]][1:5]
mk100[[9]][11:15]<-k100Dp3[[9]][1:5]
mk100[[10]][1:5]<-k100Dp1[[10]][1:5]
mk100[[10]][6:10]<-k100Dp2[[10]][1:5]
mk100[[10]][11:15]<-k100Dp3[[10]][1:5]
mk100[1:7]<-k100IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn3e5/k20IXs.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k20Dp1.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k20Dp2.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k20Dp3.RData")
mk20<-vector("list",10)
mk20[[8]]<-rbind(k20Dp1[[8]],k20Dp2[[8]],k20Dp3[[8]])
mk20[[9]][1:5]<-k20Dp1[[9]][1:5]
mk20[[9]][6:10]<-k20Dp2[[9]][1:5]
mk20[[9]][11:15]<-k20Dp3[[9]][1:5]
mk20[[10]][1:5]<-k20Dp1[[10]][1:5]
mk20[[10]][6:10]<-k20Dp2[[10]][1:5]
mk20[[10]][11:15]<-k20Dp3[[10]][1:5]
mk20[1:7]<-k20IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn3e5/k10.RData")
mk10<-k10
load("/Users/solson/Rprojects/recce1/simn3e5/k5.RData")
mk5<-k5
load("/Users/solson/Rprojects/recce1/simn3e5/k3IXs.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k3Dp1.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k3Dp2.RData")
load("/Users/solson/Rprojects/recce1/simn3e5/k3Dp3.RData")
mk3<-vector("list",10)
mk3[[8]]<-rbind(k3Dp1[[8]],k3Dp2[[8]],k3Dp3[[8]])
mk3[[9]][1:5]<-k3Dp1[[9]][1:5]
mk3[[9]][6:10]<-k3Dp2[[9]][1:5]
mk3[[9]][11:15]<-k3Dp3[[9]][1:5]
mk3[[10]][1:5]<-k3Dp1[[10]][1:5]
mk3[[10]][6:10]<-k3Dp2[[10]][1:5]
mk3[[10]][11:15]<-k3Dp3[[10]][1:5]
mk3[1:7]<-k3IXs[1:7]
rm(k100IXs,k100Dp1,k100Dp2,k100Dp3,k20IXs,k20Dp1,k20Dp2,k20Dp3,k10,k5,k3IXs,k3Dp1,k3Dp2,k3Dp3)



load("/Users/solson/Rprojects/recce1/simn1e5/k100IXs.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k100dp1.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k100dp2.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k100dp3.RData")
lk100<-vector("list",10)
lk100[[8]]<-rbind(k100dp1[[8]],k100dp2[[8]],k100dp3[[8]])
lk100[[9]][1:5]<-k100dp1[[9]][1:5]
lk100[[9]][6:10]<-k100dp2[[9]][1:5]
lk100[[9]][11:15]<-k100dp3[[9]][1:5]
lk100[[10]][1:5]<-k100dp1[[10]][1:5]
lk100[[10]][6:10]<-k100dp2[[10]][1:5]
lk100[[10]][11:15]<-k100dp3[[10]][1:5]
lk100[1:7]<-k100IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn1e5/k20IXs.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k20dp1.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k20dp2.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k20dp3.RData")
lk20<-vector("list",10)
lk20[[8]]<-rbind(k20dp1[[8]],k20dp2[[8]],k20dp3[[8]])
lk20[[9]][1:5]<-k20dp1[[9]][1:5]
lk20[[9]][6:10]<-k20dp2[[9]][1:5]
lk20[[9]][11:15]<-k20dp3[[9]][1:5]
lk20[[10]][1:5]<-k20dp1[[10]][1:5]
lk20[[10]][6:10]<-k20dp2[[10]][1:5]
lk20[[10]][11:15]<-k20dp3[[10]][1:5]
lk20[1:7]<-k20IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn1e5/k10IXs.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k10dp1.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k10dp2.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k10dp3.RData")
lk10<-vector("list",10)
lk10[[8]]<-rbind(k10dp1[[8]],k10dp2[[8]],k10dp3[[8]])
lk10[[9]][1:5]<-k10dp1[[9]][1:5]
lk10[[9]][6:10]<-k10dp2[[9]][1:5]
lk10[[9]][11:15]<-k10dp3[[9]][1:5]
lk10[[10]][1:5]<-k10dp1[[10]][1:5]
lk10[[10]][6:10]<-k10dp2[[10]][1:5]
lk10[[10]][11:15]<-k10dp3[[10]][1:5]
lk10[1:7]<-k10IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn1e5/k5IXs.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k5dp1.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k5dp2.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k5dp3.RData")
lk5<-vector("list",10)
lk5[[8]]<-rbind(k5dp1[[8]],k5dp2[[8]],k5dp3[[8]])
lk5[[9]][1:5]<-k5dp1[[9]][1:5]
lk5[[9]][6:10]<-k5dp2[[9]][1:5]
lk5[[9]][11:15]<-k5dp3[[9]][1:5]
lk5[[10]][1:5]<-k5dp1[[10]][1:5]
lk5[[10]][6:10]<-k5dp2[[10]][1:5]
lk5[[10]][11:15]<-k5dp3[[10]][1:5]
lk5[1:7]<-k5IXs[1:7]
load("/Users/solson/Rprojects/recce1/simn1e5/k3IXs.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k3dp1.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k3dp2.RData")
load("/Users/solson/Rprojects/recce1/simn1e5/k3dp3.RData")
lk3<-vector("list",10)
lk3[[8]]<-rbind(k3dp1[[8]],k3dp2[[8]],k3dp3[[8]])
lk3[[9]][1:5]<-k3dp1[[9]][1:5]
lk3[[9]][6:10]<-k3dp2[[9]][1:5]
lk3[[9]][11:15]<-k3dp3[[9]][1:5]
lk3[[10]][1:5]<-k3dp1[[10]][1:5]
lk3[[10]][6:10]<-k3dp2[[10]][1:5]
lk3[[10]][11:15]<-k3dp3[[10]][1:5]
lk3[1:7]<-k3IXs[1:7]


rm(k100IXs,k100Dp1,k100Dp2,k100Dp3,k20IXs,k20Dp1,k20Dp2,k20Dp3,k10IXs,k10Dp1,k10Dp2,k10Dp3,k5IXs,k5dp1,k5dp2,k5dp3,
   k3IXs,k3Dp1,k3Dp2,k3Dp3)

save(hk100,hk20,hk10,hk5,hk3,mk100,mk20,mk10,mk5,mk3,lk100,lk20,lk10,lk5,lk3,file="/Users/solson/Rprojects/recce1/out-6.Rdata")
                                        #load(file="/Users/solson/Rprojects/recce1/out-6.Rdata")
